Style[pojavexec]: move hooks to separate dirs, use logging macros
