+++ app_pojavlauncher/src/main/jni/Android.mk
@@ -28,12 +28,13 @@ LOCAL_SRC_FILES := \
     ctxbridges/osmesa_loader.c \
     ctxbridges/swap_interval_no_egl.c \
     environ/environ.c \
+    jvm_hooks/emui_iterator_fix_hook.c \
+    jvm_hooks/java_exec_hooks.c \
+    jvm_hooks/lwjgl_dlopen_hook.c \
     input_bridge_v3.c \
     jre_launcher.c \
     utils.c \
     stdio_is.c \
-    java_exec_hooks.c \
-    lwjgl_dlopen_hook.c \
     driver_helper/nsbypass.c
 
 ifeq ($(TARGET_ARCH_ABI),arm64-v8a)
@@ -45,7 +46,9 @@ include $(CLEAR_VARS)
 LOCAL_MODULE := exithook
 LOCAL_LDLIBS := -ldl -llog
 LOCAL_SHARED_LIBRARIES := bytehook pojavexec
-LOCAL_SRC_FILES := exit_hook.c
+LOCAL_SRC_FILES := \
+    native_hooks/exit_hook.c \
+    native_hooks/chmod_hook.c
 include $(BUILD_SHARED_LIBRARY)
 
 #ifeq ($(TARGET_ARCH_ABI),arm64-v8a)
