+++ app_pojavlauncher/src/main/jni/environ/environ.c
@@ -7,11 +7,14 @@
 #include <assert.h>
 #include <string.h>
 #include "environ.h"
+#define TAG __FILE_NAME__
+#include <log.h>
+
 struct pojav_environ_s *pojav_environ;
 __attribute__((constructor)) void env_init() {
     char* strptr_env = getenv("POJAV_ENVIRON");
     if(strptr_env == NULL) {
-        __android_log_print(ANDROID_LOG_INFO, "Environ", "No environ found, creating...");
+        LOGI("No environ found, creating...");
         pojav_environ = malloc(sizeof(struct pojav_environ_s));
         assert(pojav_environ);
         memset(pojav_environ, 0 , sizeof(struct pojav_environ_s));
@@ -19,8 +22,8 @@ __attribute__((constructor)) void env_init() {
         setenv("POJAV_ENVIRON", strptr_env, 1);
         free(strptr_env);
     }else{
-        __android_log_print(ANDROID_LOG_INFO, "Environ", "Found existing environ: %s", strptr_env);
+        LOGI("Found existing environ: %s", strptr_env);
         pojav_environ = (void*) strtoul(strptr_env, NULL, 0x10);
     }
-    __android_log_print(ANDROID_LOG_INFO, "Environ", "%p", pojav_environ);
+    LOGI("%p", pojav_environ);
 }
