+++ app_pojavlauncher/src/main/jni/jvm_hooks/emui_iterator_fix_hook.c
@@ -1,0 +1,45 @@
+//
+// Created by maks on 23.01.2025.
+//
+
+#include "jvm_hooks.h"
+#include <stdlib.h>
+
+#define TAG __FILE_NAME__
+#include <log.h>
+
+/**
+ * This function is meant as a substitute for SharedLibraryUtil.getLibraryPath() that just returns 0
+ * (thus making the parent Java function return null). This is done to avoid using the LWJGL's default function,
+ * which will hang the crappy EMUI linker by dlopen()ing inside of dl_iterate_phdr().
+ * @return 0, to make the parent Java function return null immediately.
+ * For reference: https://github.com/PojavLauncherTeam/lwjgl3/blob/fix_huawei_hang/modules/lwjgl/core/src/main/java/org/lwjgl/system/SharedLibraryUtil.java
+ */
+jint getLibraryPath_fix(__attribute__((unused)) JNIEnv *env,
+                        __attribute__((unused)) jclass class,
+                        __attribute__((unused)) jlong pLibAddress,
+                        __attribute__((unused)) jlong sOutAddress,
+                        __attribute__((unused)) jint bufSize){
+    return 0;
+}
+
+/**
+ * Install the linker hang mitigation that is meant to prevent linker hangs on old EMUI firmware.
+ */
+void installEMUIIteratorMititgation(JNIEnv *env) {
+    if(getenv("POJAV_EMUI_ITERATOR_MITIGATE") == NULL) return;
+    LOGI("Installing...");
+    jclass sharedLibraryUtil = (*env)->FindClass(env, "org/lwjgl/system/SharedLibraryUtil");
+    if(sharedLibraryUtil == NULL) {
+        LOGE("Failed to find target class");
+        (*env)->ExceptionClear(env);
+        return;
+    }
+    JNINativeMethod getLibraryPathMethod[] = {
+            {"getLibraryPath", "(JJI)I", &getLibraryPath_fix}
+    };
+    if((*env)->RegisterNatives(env, sharedLibraryUtil, getLibraryPathMethod, 1) != 0) {
+        LOGE("Failed to register the mitigation method");
+        (*env)->ExceptionClear(env);
+    }
+}
