+++ app_pojavlauncher/src/main/jni/log.h
@@ -1,6 +1,9 @@
-#ifdef __ANDROID__
+#ifndef POJAVLAUNCHER_LOG_H
+#define POJAVLAUNCHER_LOG_H
+
 #include <android/log.h>
 
+#ifndef TAG
 #define TAG "jrelog"
 #endif
 
@@ -8,12 +11,14 @@
 extern "C" {
 #endif
 
-#define LOGE(...) __android_log_print(ANDROID_LOG_INFO,    TAG, __VA_ARGS__)
-#define LOGW(...) __android_log_print(ANDROID_LOG_SILENT,    TAG, __VA_ARGS__)
-#define LOGI(...) __android_log_print(ANDROID_LOG_SILENT,    TAG, __VA_ARGS__)
-#define LOGD(...) __android_log_print(ANDROID_LOG_INFO,    TAG, __VA_ARGS__)
+#define LOGE(...) __android_log_print(ANDROID_LOG_ERROR,    TAG, __VA_ARGS__)
+#define LOGW(...) __android_log_print(ANDROID_LOG_WARN,    TAG, __VA_ARGS__)
+#define LOGI(...) __android_log_print(ANDROID_LOG_INFO,    TAG, __VA_ARGS__)
+#define LOGD(...) __android_log_print(ANDROID_LOG_DEBUG,    TAG, __VA_ARGS__)
 
 #ifdef __cplusplus
 }
+#endif
+
 #endif
